requirements.txt
line-bot-sdk
flask
appdirs       
beautifulsoup4   
bs4              
certifi          
chardet          
click            
cycler           
distlib          
filelock         
future           
graphviz         
idna             
imgurpython      
itsdangerous     
Jinja2           
joblib           
kiwisolver       
MarkupSafe      
numpy          
pandas         
Pillow           
pip              
pyparsing        
python-dateutil  
python-dotenv    
pytz             
requests        
setuptools      
six             
soupsieve        
transitions     
urllib3          
virtualenv       
virtualenv-clone 
Werkzeug 